Display new messages from server
Still having trouble with TCP socket. The server things that the pipe is broken
if and only if I start reading the connection after I initialize the TUI.
For the life of me, I can't think how those two things interact. I thought that
it was timing, but I'm less sure now.