Use reflection to detect typed nil
Once I actually wrote the code, it's obvious that this is the cleanest
way to solve the problem. I know that it has runtime overhead of some
kind, but it's not like you call this constructor in a tight loop.

If and when performance starts to matter, we can find better ways to do
this.